Refactor: Introduce EventHandler class to manage event handling logic
